# perf: squash all migrations