A prep step for upgrading to Laravel 11. ([ref](https://laravel.com/docs/11.x/upgrade#modifying-columns))

> When modifying a column, you must now explicitly include all the modifiers you want to keep on the column definition after it is changed. Any missing attributes will be dropped ... In Laravel 10, this migration would retain the unsigned, default, and comment attributes on the column. However, in Laravel 11, the migration must now also include all of the attributes that were previously defined on the column. Otherwise, they will be dropped.

> If you do not want to update all of the existing "change" migrations in your application to retain the column's existing attributes, you may simply [squash your migrations](https://laravel.com/docs/11.x/migrations#squashing-migrations).

I've confirmed that migrating a fresh database still works. All migrations in _platform_, _community_ and the migrations root folders have been deleted. I left _upcoming_ alone. In the _platform_ and _community_ folders, I added _.gitkeep_ files so the empty folders wouldn't get blown away by git.

By squashing migrations, we also manage to bring down non-parallel test execution time by roughly 33%. On my machine it decreases from 180s to 120s. Parallel execution time decreases from 30s to 21s.

Additionally, [I've added Pest](https://discord.com/channels/476211979464343552/1026595325038833725/1153294768802107452) as a dependency. Pest seems to magically take over test execution when installed -- I found during local benchmarks that this also made a ~5% improvement in test execution time.

Performance improvements are noticeable in CI as well.

Once we are on Laravel 11, we should get [another significant improvement](https://laravel.com/docs/11.x/releases#database-performance) in sqlite testing performance.

**Before**
![Screenshot 2024-08-09 at 10 24 57 AM](https://github.com/user-attachments/assets/9652cbf8-e776-43bb-81d9-f497d76387e6)


**After**
![Screenshot 2024-08-09 at 10 25 27 AM](https://github.com/user-attachments/assets/4351262a-f597-42d7-9b78-9ac35baa0dbc)