@wescopeland 2024-08-17
@Jamiras One thing I'm curious about when you review this PR: are you able to run tests in parallel mode?

`composer test -- --parallel`

I believe Pest uses a different [parallel testing](https://pestphp.com/docs/optimizing-tests) engine than PHPUnit does. I am not sure how a Windows machine will react when parallel testing mode is used, but if things are fine, we may be able to switch the pre-push hook to use parallel testing.