@Jamiras 2024-08-17
> I am not sure how a Windows machine will react when parallel testing mode is used...

I still run RAWeb in a Ubuntu VM within Windows, so I doubt I can provide the information you're looking for.