@wescopeland 2024-08-19
Sorry @Jamiras, I think I managed to confuse myself very much here.

I think the issues should be addressed now though. My intention was to also squash _2024_08_07_000000_update_emailconfirmations_table.php_. I've confirmed migrating a fresh database & tests are now all working again.